towards support for multiple filesystem instances in mfs

+++ stdinc/sys/types.h
@@ -7,7 +7,7 @@
 typedef ptrdiff_t ssize_t;
 typedef size_t off_t;
 
-typedef int dev_t;
+typedef uint64_t dev_t;
 typedef uint64_t ino_t;
 typedef short nlink_t;
 typedef short mode_t;
